FIle relocation, Renderable type enum and cmake
