+++ engine/components/CMakeLists.txt
@@ -1,5 +1,5 @@
 ﻿target_sources(${PROJECT_NAME} PRIVATE
         Component.cpp
         ComponentTransform.cpp
-        Renderable.cpp
 )
+add_subdirectory(renderables)
