+++ engine/components/renderables/CMakeLists.txt
@@ -1,0 +1,3 @@
+﻿target_sources(${PROJECT_NAME} PRIVATE
+    Renderable.cpp
+)
